wip, fixing tests
