--- Sayedha.Analyzers/Sayedha.Analyzers.Tests/packages.config
@@ -1,28 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Humanizer.Core" version="2.14.1" targetFramework="net472" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="8.0.0" targetFramework="net472" />
-  <package id="Microsoft.CodeAnalysis.Analyzers" version="3.11.0" targetFramework="net472" developmentDependency="true" />
-  <package id="Microsoft.CodeAnalysis.Common" version="4.13.0" targetFramework="net472" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="4.13.0" targetFramework="net472" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Workspaces" version="4.13.0" targetFramework="net472" />
-  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="4.13.0" targetFramework="net472" />
-  <package id="MSTest.TestAdapter" version="2.2.10" targetFramework="net472" />
-  <package id="MSTest.TestFramework" version="2.2.10" targetFramework="net472" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
-  <package id="System.Collections.Immutable" version="8.0.0" targetFramework="net472" />
-  <package id="System.Composition" version="8.0.0" targetFramework="net472" />
-  <package id="System.Composition.AttributedModel" version="8.0.0" targetFramework="net472" />
-  <package id="System.Composition.Convention" version="8.0.0" targetFramework="net472" />
-  <package id="System.Composition.Hosting" version="8.0.0" targetFramework="net472" />
-  <package id="System.Composition.Runtime" version="8.0.0" targetFramework="net472" />
-  <package id="System.Composition.TypedParts" version="8.0.0" targetFramework="net472" />
-  <package id="System.IO.Pipelines" version="8.0.0" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.5" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
-  <package id="System.Reflection.Metadata" version="8.0.0" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net472" />
-  <package id="System.Text.Encoding.CodePages" version="7.0.0" targetFramework="net472" />
-  <package id="System.Threading.Channels" version="7.0.0" targetFramework="net472" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
-</packages>
