--- Sayedha.Analyzers/Sayedha.Analyzers.Tests2/Class1.cs
@@ -1,8 +1,0 @@
-﻿public static class Program2 {
-    public static void Main2() {
-        var handler = new HttpClientHandler {
-            ServerCertificateCustomValidationCallback = (_, __, ___, ____) => true
-        };
-    }
-}
-
