+++ Sayedha.Analyzers/Sayedha.Analyzers.sln
@@ -7,13 +7,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sayedha.Analyzers", "src\Sa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sayedha.Analyzers.Shared", "Sayedha.Analyzers.Shared\Sayedha.Analyzers.Shared.csproj", "{AB791F4A-FB8A-4850-B15B-E828F22602D9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sayedha.Analyzers.Tests", "Sayedha.Analyzers.Tests\Sayedha.Analyzers.Tests.csproj", "{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sayedha.Analyzers.ProblemSamples", "samples\Sayedha.Analyzers.ProblemSamples\Sayedha.Analyzers.ProblemSamples.csproj", "{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sayedha.Analyzers.Tests2", "Sayedha.Analyzers.Tests2\Sayedha.Analyzers.Tests2.csproj", "{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sayedha.Analyzers.Tests", "Sayedha.Analyzers.Tests\Sayedha.Analyzers.Tests.csproj", "{789BF398-B485-578D-8519-25CD4C874F4C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -49,18 +47,6 @@ Global
 		{AB791F4A-FB8A-4850-B15B-E828F22602D9}.Release|arm64.Build.0 = Release|Any CPU
 		{AB791F4A-FB8A-4850-B15B-E828F22602D9}.Release|x86.ActiveCfg = Release|Any CPU
 		{AB791F4A-FB8A-4850-B15B-E828F22602D9}.Release|x86.Build.0 = Release|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Debug|arm64.ActiveCfg = Debug|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Debug|arm64.Build.0 = Debug|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Debug|x86.Build.0 = Debug|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Release|arm64.ActiveCfg = Release|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Release|arm64.Build.0 = Release|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Release|x86.ActiveCfg = Release|Any CPU
-		{2DAEAF39-5423-4642-AC37-B3566CC6ACF0}.Release|x86.Build.0 = Release|Any CPU
 		{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}.Debug|arm64.ActiveCfg = Debug|Any CPU
@@ -73,18 +59,18 @@ Global
 		{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}.Release|arm64.Build.0 = Release|Any CPU
 		{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}.Release|x86.ActiveCfg = Release|Any CPU
 		{8F433F73-F3ED-E3CE-AB53-DB368ADC9E86}.Release|x86.Build.0 = Release|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Debug|arm64.ActiveCfg = Debug|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Debug|arm64.Build.0 = Debug|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Debug|x86.Build.0 = Debug|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Release|arm64.ActiveCfg = Release|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Release|arm64.Build.0 = Release|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Release|x86.ActiveCfg = Release|Any CPU
-		{F0E95A22-7DD2-4D13-87A7-8DE7F60C3867}.Release|x86.Build.0 = Release|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Debug|arm64.Build.0 = Debug|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Debug|x86.Build.0 = Debug|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Release|arm64.ActiveCfg = Release|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Release|arm64.Build.0 = Release|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Release|x86.ActiveCfg = Release|Any CPU
+		{789BF398-B485-578D-8519-25CD4C874F4C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
